pyproject.toml
[project]
name = "rinha23"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "flask",
    "gunicorn",
    "psycopg2-binary",
    "pydantic",
]

[project.optional-dependencies]
dev = [
    "hypothesis",
    "pytest",
    "requests",
    "faker"
]
